Added new information to vote details page
